Refactoring
Updated structure to be package installable
