--- config/hpc.env
@@ -1,37 +1,0 @@
-# If you use a different High Performance Computing (HPC) system, update the environment variables below accordingly.
-
-# root project information
-ACCOUNT=pas2136
-GBIF_CACHE_ROOT=/fs/scratch/PAS2136/gbif/data
-PROCESSED_DATA_ROOT_LOCAL=/users/PAS2119/andreykopanev/gbif/processed
-PROCESSED_DATA_ROOT=/fs/scratch/PAS2136/gbif/processed
-TIME_STAMP=2024-05-01
-
-# slurm and python-slurm coordination scripts
-MPI_SUBMITTER_SCRIPT=/users/PAS2119/andreykopanev/distributed-downloader/scripts/submit_mpi_download.sh
-DOWNLOADING_SCRIPT=/users/PAS2119/andreykopanev/distributed-downloader/scripts/server_downloading.slurm
-VERIFYING_SCRIPT=/users/PAS2119/andreykopanev/distributed-downloader/scripts/server_verifying.slurm
-
-# download and log locations
-DOWNLOAD_DIR=dataset
-DOWNLOADER_URLS_FOLDER=servers_batched
-DOWNLOADER_LOGS_FOLDER=logs
-DOWNLOADER_IMAGES_FOLDER=downloaded_images
-DOWNLOADER_SCHEDULES_FOLDER=schedules
-DOWNLOADER_PROFILES_PATH=servers_profiles.csv
-DOWNLOADER_IGNORED_PATH=ignored_servers.csv
-
-# distribution of work settings
-DOWNLOADER_MAX_NODES=7
-DOWNLOADER_WORKERS_PER_NODE=6
-DOWNLOADER_CPU_PER_TASK=6
-
-# download settings
-# make sure to leave the space in the HEADER string after the colon (": ")
-HEADER="User-Agent: Imageomics Institute (https://imageomics.org; imageomics-it@osu.edu)"
-
-# IMAGE_SIZE is the max side-length of an image in pixels
-IMAGE_SIZE=720
-
-# LOGGER_LEVEL is the logging level for the downloader
-LOGGER_LEVEL=INFO
