+++ requirements.txt
@@ -40,7 +40,6 @@ pluggy==1.5.0
 psutil==5.9.8
 py4j==0.10.9.7
     # via pyspark
-pyaml==24.7.0
 pyarrow==17.0.0
 pybcj==1.0.2
 pycryptodomex==3.20.0
@@ -52,7 +51,6 @@ python-dotenv==1.0.1
 pytz==2024.1
     # via pandas
 pyyaml==6.0.1
-    # via pyaml
 pyzstd==0.16.0
 requests==2.32.3
 setuptools==71.1.0
