+++ src/distributed_downloader/tools/__init__.py
@@ -1,0 +1,1 @@
+import distributed_downloader.tools.schedulers
