--- src/tools/__init__.py
@@ -1,3 +1,0 @@
-import tools.schedulers
-import tools.runners
-import tools.filters
